重构回调相关代码并优化配置
1. 把回调和消息处理相关函数拆分到独立文件，提升代码结构清晰度
2. 优化.gitignore，增加bin文件过滤
3. 配置文件切换为性能测试模式，简化文件路径
4. 修正输出路径，便于结果统一管理
5. 注释和变量管理规范化，便于多实例和性能统计
6. 移除冗余代码，提升可维护性
