+++ src/deepstream-app/configs/yml/file_sources.csv
@@ -1,3 +1,3 @@
 enable,type,uri,num-sources,gpu-id,cudadec-memtype,rtsp-reconnect-interval-sec,rtsp-reconnect-attempts,select-rtp-protocol,smart-record,smart-rec-dir-path,smart-rec-duration,smart-rec-start-time
-1,2,file:///home/tl/data_80/data/video/110/RGB/fz_12/good/uav/Smart_Record_00005_20251209-031007_1499181.mp4,1,0,0,0,0,0,0,../smart_rec_rgb,30,3
+1,2,file:///home/nvidia/data/uav.mp4,1,0,0,0,0,0,0,../smart_rec_rgb,30,3
 0,2,file:///home/nvidia/Downloads/Smart_Record_00000_20251209-025456_1499181.mp4,1,0,0,0,0,0,0,../smart_rec_rgb,30,3
